Make reusable

+++ config/package.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-for i in base rover cons lcd ; do
-  rm ${i}.tbz
+for i in `ls -1 -d */ | sed 's%^\(.*\)/$%\1%'` ; do
+  [ -f ${i}.tbz ] && rm ${i}.tbz
   pushd $i
   tar -cvjf ../${i}.tbz .
   popd
